move ossdocker.py and format refactor
