+++ osslab/src/app/ossdocker.py
@@ -4,8 +4,18 @@ import json
 
 import requests
 
-from common import convert
 import log
+
+
+def convert(input):
+    if isinstance(input, dict):
+        return {convert(key):convert(value) for key,value in input.iteritems()}
+    elif isinstance(input, list):
+        return [convert(element) for element in input]
+    elif isinstance(input, unicode):
+        return input.encode('utf-8')
+    else:
+        return input
 
 
 def name_match(id, l):
