+++ osslab/src/app/routes.py
@@ -2,6 +2,7 @@ __author__ = "Junbo Wang"
 
 from flask import Flask, request, render_template, g ,session
 from flask.ext.restful import reqparse, abort, Api, Resource
+from time import gmtime, strftime
 import json, time, os
 from sample_course import Sample_Courses
 from expr_mgr import ExprManager;
@@ -9,6 +10,10 @@ from os.path import realpath, dirname
 from log import log
 from registration import Registration
 from database import Announcement
+from ossdocker import *
+
+start_time = strftime("%Y-%m-%d %H-%M-%S", gmtime())
+docker = OssDocker()
 
 Template_Routes = {
     "PrivacyStatement": "PrivacyStatement.html",
@@ -99,3 +104,22 @@ class DoCourse(Resource):
 class Anmt(Resource):
     def get(self):
         return Announcement.query.filter_by(enabled=1).first().json()
+
+
+class Health(Resource):
+    def get(self):
+        try:
+            health = {
+                "status": "OK",
+                "start_time": start_time,
+                "report_time": strftime("%Y-%m-%d %H-%M-%S", gmtime()),
+                "docker": docker.health()
+            }
+            return health
+        except Exception as err:
+            log.error(err)
+            return {
+                "status": "Error",
+                "start_time": start_time,
+                "report_time": strftime("%Y-%m-%d %H-%M-%S", gmtime()),
+            }
