+++ osslab/src/app/views.py
@@ -136,3 +136,4 @@ api.add_resource(CourseList, "/api/courses")
 api.add_resource(DoCourse, "/api/course/<string:id>")
 api.add_resource(StatusList, "/api/registerlist")
 api.add_resource(Anmt, "/api/announcement")
+api.add_resource(Health, "/docker/health")
